Add custom edits for half precision models
Originally intended to use the llama models but there seems to be a
difference in how the attention mask of gpt and llama models are
implemented when past_key_values is included in the decoder, causing
a size mismatch. Since I am on a deadline, it will probably suffice to use
GPT-J 6B at half precision.

+++ rome/compute_u.py
@@ -50,7 +50,7 @@ def get_inv_cov(
         )
         inv_mom2_cache[key] = torch.inverse(
             stat.mom2.moment().to("cuda")
-        ).float()  # Cast back to float32
+        ).half()  # Cast back to float32
 
     return inv_mom2_cache[key]
